Grocery Bud
Grocery List is a plain JavaScript Project without any Frameworks.  However, Google Fonts and Font Awesome will be used.  
It is similar to a to-do list.

It uses CRUD (Create, Read, Update and Delete).

User Stories
As a user, I want to add an item.
As a user, I want to remove an item.
As a user, I want to edit (update) an item.
As a user, I want to see an alert when I have added, deleted or removed 1 or more items.